+++ app/__init__.py
@@ -1,0 +1,3 @@
+from .gui import ImageConverterApp
+
+__all__ = ["ImageConverterApp"]
